new changes iteration 2

+++ model/load_images/create_directrory_for_every_file_in_directory.py
@@ -7,11 +7,12 @@ def make_directory(directory):
 
 def create_directory_for_every_file_of_directory(folder = "pictograms_test"):
     for file in os.listdir(folder):
-        folder_name = file.split("-")[0]
-        make_directory(os.path.join(folder, folder_name))
-        os.rename(os.path.join(folder, file), os.path.join(folder, os.path.join(folder_name, file)))
+        if os.path.isfile(os.path.join(folder, file)):
+            folder_name = file.split("-")[0]
+            make_directory(os.path.join(folder, folder_name))
+            os.rename(os.path.join(folder, file), os.path.join(folder, os.path.join(folder_name, file)))
 
-        #os.removedirs(file.split("-")[0])
+            #os.removedirs(file.split("-")[0])
 
 
-create_directory_for_every_file_of_directory()
+create_directory_for_every_file_of_directory("pictograms_test")
